requirements.txt
py-starbound==1.0.0